+++ pydoe/factorial/__init__.py
@@ -1,3 +1,4 @@
+from .blocking import block_full_factorial
 from .factorial import (
     alias_vector_indices,
     ff2n,
@@ -10,11 +11,17 @@ from .factorial import (
 from .fold import fold
 from .gsd import gsd
 from .john_three_quarter import john_three_quarter_design
+from .latin_square import (
+    graeco_latin_square,
+    hyper_graeco_latin_square,
+    latin_square,
+)
 from .plackett_burman import pbdesign
 
 
 __all__ = [
     "alias_vector_indices",
+    "block_full_factorial",
     "ff2n",
     "fold",
     "fracfact",
@@ -22,7 +29,10 @@ __all__ = [
     "fracfact_by_res",
     "fracfact_opt",
     "fullfact",
+    "graeco_latin_square",
     "gsd",
+    "hyper_graeco_latin_square",
     "john_three_quarter_design",
+    "latin_square",
     "pbdesign",
 ]