+++ pydoe/mixture/__init__.py
@@ -1,4 +1,13 @@
+from .axial import mixture_axial_design
+from .extreme_vertices import extreme_vertices_design
+from .process import mixture_process_design
 from .simplex import simplex_centroid_design, simplex_lattice_design
 
 
-__all__ = ["simplex_centroid_design", "simplex_lattice_design"]
+__all__ = [
+    "extreme_vertices_design",
+    "mixture_axial_design",
+    "mixture_process_design",
+    "simplex_centroid_design",
+    "simplex_lattice_design",
+]