+++ pydoe/response_surface/__init__.py
@@ -1,17 +1,21 @@
+from .blocking import block_ccdesign
 from .box_behnken import bbdesign
 from .center_design import repeat_center
 from .central_composite import ccdesign
 from .doehlert import doehlert_shell_design, doehlert_simplex_design
+from .small_composite import small_composite_design
 from .star import star
 from .union import union
 
 
 __all__ = [
     "bbdesign",
+    "block_ccdesign",
     "ccdesign",
     "doehlert_shell_design",
     "doehlert_simplex_design",
     "repeat_center",
+    "small_composite_design",
     "star",
     "union",
 ]