+++ pydoe/space_filling/stochastic/__init__.py
@@ -1,5 +1,8 @@
 from .lhs import lhs
+from .nested_lhs import nested_lhs
+from .oa_lhd import oa_lhd
 from .random_uniform import random_uniform
+from .sliced_lhs import sliced_lhs
 
 
-__all__ = ["lhs", "random_uniform"]
+__all__ = ["lhs", "nested_lhs", "oa_lhd", "random_uniform", "sliced_lhs"]