+++ pydoe/specialized/__init__.py
@@ -1,0 +1,5 @@
+from .definitive_screening import definitive_screening_design
+from .supersaturated import supersaturated_design
+
+
+__all__ = ["definitive_screening_design", "supersaturated_design"]